test: ancestor selection with multiple nestings
This test case makes sure that we can create instances of flow scopes by
referring to an indirect flow scope as the ancestor scope, even when
multiple instances of the direct flowscope already exist.

The specific examples creates multiple instances of the event sub
process, and we want to activate a task in a third instance of the event
sub process. We can achieve that by selecting the process instance as
the ancestor.