fix(tests): fix boundary event connection
When a task contains 2 boundary event, of which the 2nd one is a terminate end event we'd move the process builder to the task. If the first boundary created a gateway we should move to the gateway instead.

By adding a small check to decide which node we should move to we can make sure we end up at the correct task to continue processing as expected.