refactor(stream-platform): updates usage of tryWrite to async
Since `LogStreamWriter#tryWrite` is now asynchronous, we have to update
its usage.

There's a new flag introduced to the `ProcessingScheduleServiceImpl` to
ensure that tasks do not try to write overlapping batches. So if one
batch is currently being written in a retry loop, further batches are
buffered via the "slow" actor queue.

Use of the retry strategies is removed since those only work for
synchronous operations.