fix: clean up activateFlowScopes method
I've tried to clarify some of the steps taken by this method. This long
method is sometimes hard to wrap your head around. So I hope this helps.

Generally this commit is a refactoring, but not completely. It also
changes the behavior.

Note that I've extracted the part where we check that the ancestor
scope key is used as the direct flow scope, into a method that verifies
that the same, except it also considers the indirect flow scopes. This
leads to different behavior, for which no test was created.

I've tried to write tests that cover that case, but I've been unable to
provide it.

Lastly, this commit adds some javadoc and comments to further explain
how the method works.