test: attempt to cover indirect ancestor check
The method isAncestorOfElementInstance is used to determine whether an
element instance has a specific key as its ancestor, or in other words,
to check that it is a descendant of the ancestor scope key.

This method is able to verify this using the direct flow scope, but can
also recursively find ancestors that are the indirect flow scope.

The test case added here, tries to create a scenario where this would be
covered (it creates 4 instances of subsub, twice for each event-sub
instance). Why this does not cover the lines is a mistery to me.