fix(engine): prevent NPE when activating element inside flow scope that's being terminated
When a modify command contains an activate instruction for an element inside of a subprocess, and also a terminate instruction for the subprocess itself, Zeebe would run into a NullPointer.

This is caused by the fact that we keep track of the ancestor keys of an element that's getting activated. We need these to make sure we can clear the interrupted state from any of these ancestors. This is to ensure we support activating elements in an interrupted flow scope.

Since we first perform the activate instructions and after this the terminate instructions, we have already terminated the parent of the element that's being activated. We can not get the element instance and potentially remove the interrupted state. A null check needs to be added here.

(cherry picked from commit 8766f412bd250c8f4200bf367ba6083d87836c25)