ci(.github): switch to v1-rc1 of backport-action
This pre-release should work correctly, but I'd like to battle test it
in the Zeebe project.

This version no longer requires any additional inputs. So we can remove
any mentions of Required in the inputs. And remove any inputs that we
leave default.

We can also adjust the name of the job to the better fitting `backport`.