feat(engine): add escalation block builder
Adds an Escalation event BlockBuilder for the BPMN randomized tests. The
Escalation Event BlockBuilder generates a sub-process or a call
activity. There is a child process that throws and escalation event, and
there is a boundary escalation catch event (interupting or
non-interrupting) that catches the escalation and joins in an
appropriate gateway (exclusive or parallel).

The escalation code is unique to the generated block. At the moment,
this might be redundant, as the escalation event will always be caught
by the encapsulating boundary escalation catch event.