test(engine): verify job without retries is not activatable
When a job has failed and has no retries left, it should no longer be possible to activate this job. Activating this job would result in an incident. Since there already is an incident a second one will be created. This will conflict with the consistency checks, where we only allow a single incident for a task.

(cherry picked from commit 54b2fc3420fbd53af8e50875d3918b752abca01f)