fix(engine): prevent activation of failed job without retries
When a Job has failed and has no retries left, it should no longer be possible to activate this job. Activating this Job would result in an Incident. Since there already is an Incident a second one will be created. This will conflict with the consistency checks, where we only allow a single Incident for a task.