refactor(logstreams): remove fluent builder
Removes the log stream fluent builder interface in favor of a simple
`tryWrite` API which takes in pre-built `LogAppendEntry` (or a batch
thereof).

The writer implementations mostly perform serialization at this point,
which will be later collapsed into a single writer once the dispatcher
replacement arrives and serialization is pushed downstream.