perf(engine): lookup ancestor once
The previous code was retrieving the ancestor instance multiple times
from the state. There's no need to do this, because the ancestor doesn't
change in the meantime. So a single lookup should suffice.

Likewise, we don't have to go through the elementInstancesOfScope
multiple times to find the instance that is a flow scope of the
ancestor instance. Instead, we can store the result of the first pass
and reuse this result.