refactor: command api creates new record values for every request
Instead of using static instances of the known record values,
when handling a command api request the request reader creates a new
instance of the record value.

This ensures that handling new requests does not modify a record value
that was already written to the logstream. Currently, this is not
required as the logstream writer immediately serializes (and thus
copies) the record value. To serialize asynchronously (from the
viewpoint of the command api), the command api must write immutable
record values that do not change before they are serialized.