refactor: inter-partition command receiver writes immutable copies
When decoding messages from other partitions, buffers and metadata were
reused to decode the command and write it to the logstream. Until now,
this was fine because the writing would immediately serialize (and thus
copy) the command. Since we are planning to change this behavior to
serialize asynchronously (from the viewpoint of the inter-partition
command receiver), the decoder is now using new buffers and metadata
for every message that it decodes.