refactor: call create vars callback only once
This is the wrap up of the a series of commits to improve the
maintainability of this lone method. Specifically step 4.

Note that we only have to call the create variables callback method for
existing scopes. If we're activating a scope newly, then the callback
method is called from that part of the code to make sure the variables
are created in between the ELEMENT_ACTIVATING and ELEMENT_ACTIVATED
events.