refactor: extract findScopeKey method
The activateFlowScopes method does a few things, and does them
recursively. It becomes much easier to grasp if you don't also have to
think about 'how do find the key of the flow scope to use' in the next
recursion.

We can easily extract this part into a separate method that clearly does
1 thing: find the key of the flow scope that we want to use.