ci: create benchmarks for PRs labeled with 'benchmark'
This workflow is triggered when adding/removing a 'benchmark' label
to pull requests and when updating/closing labeled pull requests.

When adding the label or updating the pull request, a benchmark is
created or updated with a name like 'pr-1234-benchmark'. When removing
the label or closing the pull request, the underlying kubernetes
namespace is deleted.