refactor: reduce parameters to minimum
The process instance record is used in other methods for multiple
things, but in the findScopeKey method its only use is to provide the id
of the process for the feailure message in case something went wrong.

Reducing the parameters to the bare minimum is generally considered a
good practise.