docs: use subprocess over flow scope
These methods are difficult enough to understand as they are, but they
refer a lot to ancestors, and flow scopes. However, we can clarify what
we mean by this by using the word subprocess instead.

Each flow scope that is used is actually an ancestral subprocess of the
element targeted for activation.