feat(engine): catch errors without errorRef or errorCode
If 'errorRef' is omitted or 'errorCode' of referenced error is omitted,
the event start event or boundary event is triggered by any error event.

When 'errorRef' is omitted, an empty error without errorCode will be created
to facilitate finding the corresponding catch event.

When an error is thrown, we need to walk through the scope hierarchy
and look for a matching catch event. If we don't find a code-specific Error event
in the current scope, but find a catch-all Error event in that scope,
we should catch the error by the catch-all Error event.