feat(exporters/elasticsearch): store record counters as metadata
Store the record counters as metadata of the exporter by updating the exporter position. Update the position after the flushing the bulk request successfully. If the flush fails during the exporting of a record (i.e. synchronous flush) then we should not update the record counters because the exporting will be retried.

Restore the record counters from the stored metadata when opening the exporter. The metadata is stored in a JSON format.