test(logstream): don't test that appender fails on position gaps
This was testing that the appender would fail if the dispatcher would
produce gaps in the positions within one batch.
With the sequencer, the appender is in charge of serializing and
ultimately writing the positions for each entry. The sequencer is only
assigning a starting position for each batch. The test no longer makes
sense.