feat(backup): return latest checkpoint info if a checkpoint already exists
This is useful to return a useful error message to the user.

Standard used in engine is to send a rejection in such cases. But with rejection we cannot send the latest checkpointId. So we chose to send the IGNORED record. Gateway request handling will be updated to interpret this info.