fix(engine): mark flowscope of terminate end event as interrupted
The terminate end event should mark its flow scope as interrupted when it's completed. Having an interrupted flow scope means other elements in this scope can no longer be activated or completed. Any ACTIVATE/COMPLETE command will be rejected.

If we do not interrupt the flowscope it is possible for the engine write a TERMINATE command as a result of the end event. It could then go on to process a COMPLETE command. This will not be rejected, and it will take outgoing sequence flows for this element. The TERMINATE will then be rejected as the element is no longer active. Since the flowscope now has an outgoing sequence flows it is considered to have active elements, it will not be terminated, as it should upon reaching a terminate end event.

(cherry picked from commit da387a5034d0e3ccffd437e184efcb689f1b6c91)