fix(engine): complete process when interrupted by terminate end event
The terminate end event will have marked the process as interrupted. When a terminate end event is reached we expect its flowscope to be completed, not terminated. Therefore, we need to check if the process was interrupted because of a terminate end event. If this is the case we should try to complete the process.

(cherry picked from commit d9b4b88c94e2832d27eb24d3e6010c76eabc20e6)