fix(engine): complete (event) sub process when interrupted by terminate end event
The terminate end event will have marked the (event) sub process as interrupted. When a terminate end event is reached we expect its flowscope to be completed, not terminated. Therefore, we need to check if the (event) sub process was interrupted because of a terminate end event. If this is the case we should complete it

(cherry picked from commit 89e6d9b47da8b34756990a72a012680f54e13195)