feat(gateway): allow configuring gRPC max message size
The gRPC maxMessageSize was not configured so the default value of 4MiB
applied.

This adds a new setting to the gateways network config that is used
when building the netty server for gRPC.

The broker's network config already has a `maxMessageSize` which,
confusingly, was not used for the network configuration but for
controlling the maximum batch size in the dispatcher. We are reusing
this setting now to also configure gRPC max message size.

This may look somewhat confusing but both dispatcher and gRPC
`maxMessageSize` are related. The dispatcher limit should not be smaller
than the gRPC `maxMessageSize`. Additionally, the placement of
`maxMessageSize` inside the network config suggests that it was always
meant to be used as a limit on the network/gRPC side too.