fix(test): verify boundary event does not get triggered when flowscope is interrupted
If we have a subprocess containing a boundary event, and there is also an event subprocess in a higher scope than this subprocess. It could occur that both the boundary event and the event subprocess get triggered at the same time. If the boundary event got triggered first we are in a scenario where we will try to activate the boundary event.
When this happens the subprocess gets terminated and the boundary event gets activated. The terminate command triggered by the event sub process gets rejected, as the sub process is already terminating.

Because the event sub process never tries to terminate the boundary event, which was not active at the time of processing the interruption, the process instance gets stuck. The boundary event will be active and as a result the start event inside the event sub process never gets activated.

This behavior is unexpected. The event sub process should get terminated as expected and the event sub process should start processing.