test: allow ancestor selection inside multi-instance
Apart from the 2 cases described in the previous commit (both resulting
in a newly activated inner multi-instance), we should allow activations
inside multi-instance where the selected ancestor is a descendant of the
multi-instance body.

For example, it should be allowed to select the inner instance of the
multi-instance sub-process as ancestor. This would only result in the
activation of the selected element, and NOT in the new activation of an
inner instance of a multi-instance body.

The provided example was originally provided by Hüsna, where Operate
wants to 'move a token' inside a multi-instance.

It should also be possible to 'move a token' from one inner
multi-instance to another inner multi-instance. However, there is no
specific edge cases being tested there, so I haven't covered it in a
test.

We could consider adding a test case where multiple flow-nodes are
marked as multi-instance (e.g. nested multi-instance sub-processes).
Again the same rule should apply: activation instructions should not
result in the direct activation of multi-instance bodies, or of the
inner instance of multi-instances.