feat(test): re-enable escalation events in process generation
This blockbuilder was disabled as it resulted in a lot of failing builds. The cause for this were multiple events triggering at nearly the same time, conflicting with eachother. The introduction of the escalation blockbuilder made this a lot more likely to happen.

Since this bug is fixed we can re-enable this blockbuilder.