fix(engine): don't trigger boundary event if flowscope is already interrupting
If a flowscope is already interrupted we should not trigger new events. Doing this would result in the activation of elements which the original interrupting event did not consider. This means there are elements active, which will not be terminated as a result of the interrupting event, causing the process to become stuck