fix(gateway): omit failurereason from list response if status is not failed
sbe response contains and empty string, which was then considered as non-null when building the PartitionStatus response. Since it is not null, it is also serialized as empty string. On the client side if we print the response of list, then it unnecessarily also include the failurereason. This is fixed in this commit.