fix: reject activate with multi-instance between ancestor
Okay okay, perhaps a bit of a cryptic commit summary, but the gist is
this:

We want to reject modifications with activate instructions that try to
activate an element using ancestor selection where a multi-instance
element lies inbetween the element targetted for activation and the
selected ancestor.

We cannot support this at this time, because it would require us to
activate that multi-instance, including one of its inner instances. This
is hard, because we'd have to deal with the loopCounter, input element,
and other multi-instance properties.

This case actually is part 2 of cases that would result in us having to
activate a multi-instance. For the previous work, see the parent of the
parent of this commit.