fix: detect when problem occurs instead of earlier
This code is mostly a refactoring of the logic where we reject an
activate instruction because it attempts to activate a multi-instance
body or an inner instance of a multi-instance body, as an ancestral flow
scope of the target element.

The previous code tried to check whether the activate instruction would
lead to a specific situation (i.e. it would validate the command). This
required a lot of additional code.

We only need these validation because we do not yet support this case,
but we plan to do so in the future. Therefore, I think it's better to
not be too specific and simply fail when we can't support the action.

To do that, we can simply throw a new
UnsupportedMultiInstanceBodyActivationException and handle it as an
expected error in the ProcessInstanceModificationProcessor.

This requires much less code, and is much easier to replace when we
eventually implement this complicated part of the feature.

Note that the new exception, leads to a different error message. So the
behavior is different from the user's perspective.