feat(engine): reject modification if ancestor does not exist or is not active
When a modification contains an activate instruction with an ancestor scope key, we must have an element instance for this ancestor in our state. This element instance must be in an active state. If this is not the case we cannot activate an element inside of this scope and the modification should be rejected.