test(engine): verify ancestor key belonging to different process instance is rejected
When activating an element using modification the user has the option to provide an ancestor key. There is no guarantee that this key is correct and that it belongs to the process instance that is being modified. We need to make sure that if the key does not belong to the modified process instance, we reject the command.