chore(.github): add backport workflow (#5649)
This adds a workflow to automatically backport pull request once they
are merged to maintenance branches.

The referred PAT `BACKPORT_ACTION_PAT` is a Personal Access Token of
@backport-action (a GH bot user controlled by @korthout). The PAT will
expire in May 2023, but @korthout will refresh the token in time.

If you ever wish to replace the PAT with another user', then make sure
that user has write access to the repo.

There are other things that can be configured for this action, but this
seems like a good starting point. Please refer to
https://github.com/zeebe-io/backport-action for a full list of the
available inputs for configuration.

I hope this workflow solves a problem for you. Let me know if it causes
any. :D