feat(engine): support catch-all and code-specific escalation catch events
As a user, in a given scope, I can define:

* A single Escalation catch event with no escalation code defined (a catch-all Escalation event).
* Multiple Escalation catch events with a defined escalation code.

When an escalation is thrown, we need to walk through the scope hierarchy
and look for a matching catch event. If we don't find a code-specific escalation event
in the current scope, but find a catch-all Escalation event in that scope,
we should catch the escalation by the catch-all Escalation event.