fix(engine): backport merge conflict fixes
In the backported code we have a BpmnBehavior object available in most of the processors. This is a refactoring that was not yet done in version 8.0. Therefore, this commit fixes the compilation errors that were here after the cherry-pick. It creates a BpmnStateBehavior and passes it to the processors that need it.