feat(engine): add process instance key property to EventTrigger
A process could contain multiple start events. Depending on whether we can find an EventTrigger we activate the triggered start event, or the none start event. If we want to activate the none start event, but we already have an EventTrigger in the state we will trigger the wrong one.

Before activation of the start event a process instance key is generated. We can store this in the EventTrigger. This can now be used in the ProcessProcessor to determine the correct start event to trigger.