feat(engine): determine which start event to trigger using process instance key
A process could contain multiple start events. Depending on whether we can find an EventTrigger we activate the triggered start event, or the none start event. If we want to activate the none start event, but we already have an EventTrigger in the state we will trigger the wrong one.

By using the process instance key in the EventTrigger we can determine whether this activation relates to the EventTrigger in the state. If this is the case we should activate this triggered start event. Otherwise we should activate the none start event.

For backwards compatibility's sake, if the process instance key is not available in the EventTrigger, we should still activate the triggered start event. This way behavior does not change.