test(engine): verify correct start event gets triggered
A process could contain multiple start events. Depending on whether we can find an EventTrigger we activate the triggered start event, or the none start event. If we want to activate the none start event, and we already have an EventTrigger in the state we have to make sure we don't activate the wrong start event.