feat(gateway): add gateway api for standalone decision evaluation
Expose the new standalone decision evaluation feature of the workflow engine in the gateway API.

For this purpose, additional rpc messages were defined. The new rpc messages mirror the DecisionEvaluationRecord, EvaluatedDecisionRecord, EvaluatedInputRecord, EvaluatedOutputRecord, MatchedRuleRecord. The goal is to provide the full data of an evaluated decision to the user. With the full data, users can obtain the final result of the evaluation, but can see any additional decision evaluations that were performed within the same decision evaluation execution.