feat(engine): subscribe processes with top-level signal start events to signals
When a new process with top-level signal start events is deployed, we need to subscribe that process to signals.

The engine can create subscriptions for signals using a new SignalSubscription record. This SignalSubscription must be CREATED when the process is deployed. If that process is a new version of an already existing process, this SignalSubscription must be DELETED for that older version of the process.

We will only subscribe to top-level signal start events in the Deployment Partition (i.e. partitionId: 1). It may be possible to also subscribe on other partitions, but that is a stretch-goal as we are not yet sure whether this has further complications.

Previously existing processes should not be able to contain top-level signal start events. There is no need to process already deployed processes additionally.