fix: reject when ancestor is not flow scope
When an ActivateInstruction of a ProcessInstanceModification command
refers to an ancestor instance (with the ancestorScopeKey) that is not
an ancestor of the element to activate (i.e. it is not a direct flow
scope and not an indirect flow scope of the element), then the command
must be rejected.

The provided rejection message contains a message about each of the
instructions that did not pass this validation.

The rejection message describes both the key (used in the instruction),
as well as the element id that this key refers to, and the element id
that they tried to activate, to clarify to the user what they attempted.
This should help them figure out what went wrong and how to resolve the
problem.