fix: adjust rejection msg to suggest ancestor selection
Originally, this part of the rejection message was left out, because
ancestor selection was not yet available. Now that an ancestor can be
selected, we should inform users that they can resolve this specific
problem by selecting an ancestor.

This commit contains both the test change and the code change in one,
because it's such a small change and easy to understand.