fix: close channels on timeout connection
We have seen in several experiments that channels have been reused even after a long period of unavailabilty of the other side. Meaning
the brokers were already restarted and had new IP's but gateways still have send requests to old IP's without realizing the issue.

This should be resolved with creating new channels when timing out, which is likely an indication of an connectivity issue.