allow overriding the authority
this is required to enable SNI and use self signed certificate, or
configure nodes with IPs instead of domains and still use TLS. it's not
great though, as you are globally overriding the authority for all
outgoing requests, meaning to all nodes.

we could either:

a) map authorities to nodes, so we can pick the right one based on the
address
b) disable it if you're trying to connect to an IP and not a domain name

I think it requires some more digging into. For backwards compat though
we could simply disable it for now as it's not enabled in the current
codebase