feat(clients/java): define new evaluateDecision command
Add a new EvaluateDecisionCommand that will implement the new evaluateDecision rpc method. The command adds support for triggering standalone decision evaluation in the workflow engine. In the first iteration of this command, users will be able to provide a decision ID, to evaluate the latest version of the decision. Otherwise, a user can provide a decision key, to evaluate a specific (older) version of a given decision. Furthermore, a user can add variables to the command to be included in the decision evaluation. The variables can be provided in multiple formats: Map, Object, InputStream, or String.