fix: support ValDate feel expression results
When using `today()`, the expression results in a ValDate, i.e. a Date
without a Time component.

When this is the result of an expression, Zeebe should be able to
recognize the result type.

Typically, the result type is checked in a switch statement. If the
switch statement doesn't cover the case, the type is considered
unexpected.

However, if there is no type, then `case _ => null` means that the value
`null` is used in the switch statement, which leads to an NPE.

Note that we can safely add ne entries to the ResultType enum as it is
only used at runtime.