feat(raft): only sync data for metastore file updates
Meta file update is execute in the hot path of raft, and have a significant influence in performance. By switching to DSYNC, we observed a tiny bit of improvement to p50 latency. Using DSYNC is safe here as the file metadata is not relevant for us.

(cherry picked from commit 5d1a381e17401fd5a6dcf940aa325e54d907b266)