fix: start the start event at y 100
The Y event is calculated by a base value + the lowest height of all elements. This makes it possible to add multiple processes to a single bpmn file, without the elements overlapping.

Since a start event 36x36 pixels we need to subtract 36 from the y coordinate. 100 - 36 = 64. We need to do this here because the start event is already added to the model. This means when we fetch the lowest Y coordinate we already get the height of the start event, which is 36.