feat: write messaging correlation directly
When message correlation is happening on the same partition, we can
directly write a follow up command instead of sending an inter partition
request.

This commit adds some conditional checks in order to distingush
between correlating to different partition or to the same partition.

In combination with batch processing this means that we can also
directly process the follow up command in the same batch and continue
with the process instance execution. This allows to remove several
commit latecny wait times.

Before this command has been send out, over loopback it will be written
to the log. It needed to be committed and read again by the processor
in order to continue. Benchmarks shows that it can help significantly on
certain user bases, which controll the message distributions and
correlations.