fix(atomix): don't step down when rebalancing a balanced cluster
Previously we were comparing the object identity of the partition's
primary and our member id to decide if we should step down or not.
This was wrong, what we actually want to compare are the values of
`MemberId` and not the object identity.

This fixes a bug where repeated rebalancing of an already balanced
cluster would introduce short availability interruptions and needless
transitions from leader to follower and back to leader again.

(cherry picked from commit d46ea0ce244fce92439b80759035da4972f69bcb)