feat(broker): extend the SBE message for exporter state
Add a new field for the exporter metadata to the SBE message for distributing the exporter state. Rename the SBE message to reflect that it doesn't contain only the exporter positions but also the metadata.

Follow the SBE version guidelines and increase the version of the SBE protocol. Previous Zeebe versions doesn't contain the new field.

The renaming of the SBE message and the fields doesn't violate the backward compatibility because SBE serializes only the ids not the names.