feat(atomix): reduce likelihood of forced snapshot replication
After taking snapshots, we'd immediately compact the log up to the
snapshot index. This would force snapshot replication for followers
which are lagging behind the leader because of slow replication. If the
leader compacted the log and doesn't have the entries needed for
replication, sending a snapshot is the only option.

With this change, the leader can more often continue with regular
replication for slow followers because it doesn't compact *everything*
up to the  snapshot index, but leaves a configurable amount of entries
in the log.
The amount of entries is based on `preferSnapshotReplicationThreshold`,
which is already used by leaders to decide between normal replication
and sending snapshots if both are possible.

(cherry picked from commit ef6203b1adc4a8d98dbf2e478368abf800a5b420)