feat(benchmark): add year to weekly medic benchmark name
The medic benchmark workflow deletes the oldest benchmark. It does this by sorting the name of benchmarks and picking the first ones. When we transition to the next year this means we consider the benchmark from week 1 to be older than the benchmark from week 51. Because of this we delete the wrong benchmark.

Introducing the year in benchmark name is an easy way to circumvent this problem.