refactor(broker): refactor snapshot director for readability
The chain of actions required to take a snapshot is difficult to understand because of the inability to compose actor futures. In this PR, we refactored it with a rudimentary way of composing actor futures.

There is no behavior change.

(cherry picked from commit 3a60c07ff7f0a1c813fe135a3f1b3c904f915dc7)