fix: retry processing with fewer commands when batch result is too large
This is incredibly hacky but should work. When a batch is being
processed and the resulting entries exceed the maximum record size,
we discard the transient state (transaction and pending writes) and
retry processing from the start, only this time limited so that only
the accepted number of commands will be processed at once.

When processing with a command count limit, all commands that don't fit
within that limit are written with a magic source index: -2.
When serializing such a command, we don't set the source position to the
initial command (when source index is -1) or to the position of another
record in the batch (when source index is > 0) but instead set the
source position to -1.
For log readers, this command appears like an external command. During
processing, we don't recognize this command as already processed and
thus skip the command.

This approach is not ideal. Besides the hacky implementation, the main
drawback is that we lose association of unprocessed commands to the
record which caused them.