fix(engine): prevent NPE when process instance is terminated
Users can pass the process instance key as a terminate instruction in a modification command. If they do this, and also have an activate instruction we will run in a NPE during processing. This is because we try to find the process instance during the processing of the MODIFIED event. Since the process instance is terminate as a result of the instructions, this no longer exists.

We always execute the activate instructions first. The expected result in this situation would be sending the activate command of the activate instruction and then terminating the process instance. This will result in a terminated process instance and a rejected activate command.