refactor(engine): adjust decision id or key check on evaluation
When evaluating a decision, the processor performs a check if a decision ID or decision KEY is provided. Previously, the check performed an XNOR on the ID and KEY values. If both or none were provided, a Failure was returned. This way of checking for ID and KEY values doesn't work well for all Zeebe clients. In the Go client, the default value for the KEY (long) is zero, which is a valid KEY. Overriding the default decision KEY value to -1 in the Go client isn't straightforward. As a result, it was easier to refactor the checking logic to give priority to the decision ID. The same pattern is applied in the CreateProcessInstanceProcessor.